Fixed Source Port List referencing issues
Took me awhile to understand this one and to find the best way to correct this issue.

Fixed a critical issue where the original list would always be manipulated regardless if the user saved the list or not.  This was caused by the reference of the original list to be manipulated directly and was never treated by value but by a reference.