Added Region directives
Applied region directives for member variables.   This should make it easier to read and minimize distractions.

I will not go crazy with these directives as Visual Studio makes it a bit hard to manage with the spacing.  I don't remember it being this difficult with C#, but in VB - it's a take or leave it cases with the spacing....