Always mirror size of List<T>
Previously, if the List<T> size was greater than 0 - then we capture the List<T> size.  However, we should always capture the size regardless if any entries exists or not.  This bit of information is critically necessary when a new session has begun within this program.  When the program starts, it will (or soon) load the saved data - and thus needs the size of the List<T>'s, even if '0'.