Detect if settings exists in Registry
Check to see if the program's registry subkey root exists.  If in case the root subkey does not exist, then assume that no data is available for use.  Though, if the subkey does exist - then we assume that data exists and ready for access.

Removed 'FirstRun' key.  I guess I wont need it right now.