Reset all local user preferences
When the user requests for a reset, local data must also be reset along with the Registry values.

Previously, only the registry values was reset - but the local values was never cleared or set to a default value.  Thus, this feature was only half-way implemented and was not yet implemented until now.