WIP: Progress Update - ComboBoxes
Still working on this...

The idea thus far, as ComboBox's do NOT provide the index that is selected, we have to do that ourselves.  To accomplish this, we test the selected string and compare it to the respected list to get the index key that was selected from the list.