Horizontally resized parent window
It seems as if the PWAD Group is smashed against the Window border.  The best ultimate fix is to make this program's UI more responsive to different Window sizes, but I do not have much time left - only have less than 48 hours....  The only fix I can push is to horizontally resize the window size.